service-clients/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.2
python-dotenv==0.21.0
Werkzeug==2.2.2
cryptography
Faker==25.2.0
